chore : move enum values to entity
