--- src/constants.ts
@@ -1,9 +1,0 @@
-export enum VideoType {
-  LOCAL = 'LOCAL', // 로컬에서 업로드
-  YOUTUBE = 'YOUTUBE', // 유튜브에서 업로드
-}
-
-export enum VideoLanguage {
-  KOR = 'KOR', // 한국어 자막
-  ENG = 'ENG', // 영어 자막
-}
